# cmd/docs/swagger.yaml
definitions:
  errors.Error:
    properties:
      errors:
        additionalProperties: true
        type: object
    type: object
  handler.AccountBalanceResponse:
    properties:
      address:
        type: string
      balanceETH:
        type: string
      balanceFiat:
        type: string
    type: object
  handler.CreateAccountRequest:
    properties:
      password:
        type: string
    required:
    - password
    type: object
  handler.CreateAccountResponse:
    properties:
      address:
        type: string
    type: object
  handler.CreateBillRequest:
    properties:
      address:
        type: string
      amount:
        type: number
      password:
        type: string
      payers:
        items:
          type: string
        type: array
    required:
    - address
    - amount
    - password
    - payers
    type: object
  handler.CreateBillResponse:
    properties:
      billAddress:
        type: string
      success:
        type: boolean
      transactionHash:
        type: string
    type: object
  handler.CreateTransferRequest:
    properties:
      amount:
        type: number
      fromAddress:
        type: string
      password:
        type: string
      toAddress:
        type: string
    required:
    - amount
    - fromAddress
    - password
    - toAddress
    type: object
  handler.CreateTransferResponse:
    properties:
      success:
        type: boolean
    type: object
  handler.UserLoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  handler.UserRegisterRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  handler.UserResponse:
    properties:
      email:
        type: string
      image:
        type: string
      token:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/accounts/{address}/balance:
    get:
      consumes:
      - application/json
      description: Get the account balance
      operationId: get-account-balance
      parameters:
      - description: address of the account balance to get
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.AccountBalanceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - ApiKeyAuth: []
      summary: Get the account balance
      tags:
      - accounts
  /api/accounts/create:
    post:
      consumes:
      - application/json
      description: Creates a new account
      operationId: create-account
      parameters:
      - description: Info for creating account
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/handler.CreateAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.CreateAccountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Error'
        "404":
          description: Not Found
          schema:
            type: objects
        "500":
          description: Internal Server Error
          schema:
            type: objects
      security:
      - ApiKeyAuth: []
      summary: Creates a new account
      tags:
      - accounts
  /api/bills/create:
    post:
      consumes:
      - application/json
      description: Creates a new bill
      operationId: create-bill
      parameters:
      - description: Info for creating bill
        in: body
        name: bill
        required: true
        schema:
          $ref: '#/definitions/handler.CreateBillRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.CreateBillResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Error'
        "404":
          description: Not Found
          schema:
            type: objects
        "500":
          description: Internal Server Error
          schema:
            type: objects
      security:
      - ApiKeyAuth: []
      summary: Creates a new bill
      tags:
      - bills
  /api/transfers/make:
    post:
      consumes:
      - application/json
      description: Creates a new transfer
      operationId: create-transfer
      parameters:
      - description: Info for creating transfer
        in: body
        name: transfer
        required: true
        schema:
          $ref: '#/definitions/handler.CreateTransferRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.CreateTransferResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Error'
        "404":
          description: Not Found
          schema:
            type: objects
        "500":
          description: Internal Server Error
          schema:
            type: objects
      security:
      - ApiKeyAuth: []
      summary: Creates a new transfer
      tags:
      - transfers
  /api/user:
    get:
      consumes:
      - application/json
      description: Gets the currently logged-in user
      operationId: current-user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Error'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - ApiKeyAuth: []
      summary: Get the current user
      tags:
      - user
  /api/user/login:
    post:
      consumes:
      - application/json
      description: Login for existing user
      operationId: login
      parameters:
      - description: Credentials to use
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handler.UserLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Error'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Error'
      summary: Login for existing user
      tags:
      - user
  /api/user/signup:
    post:
      consumes:
      - application/json
      description: Register a new user
      operationId: sign-up
      parameters:
      - description: User info for registration
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handler.UserRegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Error'
        "404":
          description: Not Found
          schema:
            type: objects
        "500":
          description: Internal Server Error
          schema:
            type: objects
      summary: Register a new user
      tags:
      - user
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"